Moved ver rpc server to rpcservers folder
